Use web3j gradle plugin for acceptance tests
Automatically generate wrappers for all contracts in
acceptance-tests/tests/contracts
