+++ acceptance-tests/tests/build.gradle
@@ -11,6 +11,21 @@
  * specific language governing permissions and limitations under the License.
  */
 
+plugins {
+  id 'org.web3j' version '4.8.4'
+  id 'org.web3j.solidity' version '0.3.2'
+}
+
+web3j {
+  generatedPackageName = 'org.hyperledger.besu.tests.web3j.generated'
+}
+
+sourceSets.main.solidity.srcDirs = ["$projectDir/contracts"]
+
+solidity {
+  resolvePackages = false
+}
+
 dependencies {
   implementation project(':crypto')
