+++ acceptance-tests/tests/contracts/CrossContractReader.sol
