+++ acceptance-tests/tests/contracts/SimpleStorage.sol
