+++ acceptance-tests/tests/src/test/java/org/hyperledger/besu/tests/acceptance/privacy/EthSignerAcceptanceTest.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.tests.web3j.privacy;
+package org.hyperledger.besu.tests.acceptance.privacy;
 
 import static org.web3j.utils.Restriction.UNRESTRICTED;
